Implement `ChatClient`

+++ Data/GetChatsRequest.cs
@@ -5,6 +5,6 @@
 [Serializable]
 public record GetChatsRequest : AuthedRequestBase {
 	public required string[] usernames { get; set; }
-	public float? before { get; set; }
-	public float? after { get; set; }
+	public double? before { get; set; }
+	public double? after { get; set; }
 }
